Refactor Navbar and Sidebar components
