+++ src/components/shared/navbar/NavbarComponent.tsx
@@ -2,7 +2,7 @@ import { Flip } from '@mui/icons-material';
 import { IconButton, Toolbar, Typography } from '@mui/material';
 
 import { useLocation } from 'react-router-dom';
-import { AppBar } from './NavbarStyle';
+import { AppBar } from './NavbarStyles';
 
 export const NavbarComponent = ({
   setOpen,
