+++ src/components/shared/navbar/NavbarStyles.tsx
