+++ src/components/shared/sidebar/SidebarStyles.tsx
@@ -1,0 +1,51 @@
+import { styled, Theme, CSSObject } from '@mui/material/styles';
+import MuiDrawer from '@mui/material/Drawer';
+import { blueTheme } from '../../../theme/blueTheme';
+import { drawerWidth } from '../constants';
+
+const openedMixin = (theme: Theme = blueTheme): CSSObject => ({
+  width: drawerWidth,
+  backgroundColor: theme.palette.primary.light,
+  color: theme.palette.getContrastText(theme.palette.primary.light),
+  height: `calc(100% - ${theme.spacing(4)})`,
+  borderRadius: '24px 24px 24px 24px',
+  transition: theme.transitions.create('width', {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.enteringScreen,
+  }),
+  overflowX: 'hidden',
+});
+
+const closedMixin = (theme: Theme = blueTheme): CSSObject => ({
+  backgroundColor: theme.palette.primary.light,
+  color: theme.palette.getContrastText(theme.palette.primary.light),
+  height: `calc(100% - ${theme.spacing(3)})`,
+  borderRadius: '24px 24px 24px 24px',
+  transition: theme.transitions.create('width', {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.leavingScreen,
+  }),
+  overflowX: 'hidden',
+  width: `calc(${theme.spacing(7)} + 1px)`,
+  [theme.breakpoints.up('sm')]: {
+    width: `calc(${theme.spacing(8)} + 1px)`,
+  },
+});
+
+export const Drawer = styled(MuiDrawer, {
+  shouldForwardProp: (prop) => prop !== 'open',
+})(({ theme = blueTheme, open }) => ({
+  color: theme.palette.getContrastText(theme.palette.warning.light),
+  width: drawerWidth,
+  flexShrink: 0,
+  whiteSpace: 'nowrap',
+  boxSizing: 'border-box',
+  ...(open && {
+    ...openedMixin(theme),
+    '& .MuiDrawer-paper': openedMixin(theme),
+  }),
+  ...(!open && {
+    ...closedMixin(theme),
+    '& .MuiDrawer-paper': closedMixin(theme),
+  }),
+}));
